refactor: SectionEditor.tsxを分割してコードの保守性を向上
- SectionEditor.tsx (572行) を4つのファイルに分割
- SectionCard.tsx: 個別セクションの描画ロジック
- useSectionOperations.ts: セクション操作のビジネスロジック
- useChordOperations.ts: コード操作のビジネスロジック
- SectionEditor.tsx: メインコンポーネント (175行に短縮)

改善点:
- 責任の分離により可読性が向上
- カスタムフックの再利用性向上
- 小さなファイルによりテストが容易
- 保守性の大幅な改善
